Build fix.
Change set 114220 broke OS X internal build. Change set 114231 attempted to fix it, but isn't entirely
        correct. This patch amends change set 114231.

* WebCore.exp.in:
* editing/DictationAlternative.cpp:
(WebCore::DictationAlternative::DictationAlternative):
* editing/DictationAlternative.h:
(DictationAlternative):

Canonical link: https://commits.webkit.org/101541@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114266 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/editing/DictationAlternative.cpp
@@ -28,7 +28,7 @@
 
 namespace WebCore {
 
-DictationAlternative::DictationAlternative(unsigned start, unsigned length, unsigned context)
+DictationAlternative::DictationAlternative(unsigned start, unsigned length, uint64_t context)
     : rangeStart(start)
     , rangeLength(length)
     , dictationContext(context)
